IT DOESNT BREAK OMG

+++ src/actions/actions.js
@@ -117,18 +117,22 @@ function deletePlant(plant){
   console.log(plant.id);
 
   return dispatch => {
-    fetch(`http://localhost:3000/api/v1/plants/${plant.id}`, {
-      method: 'DELETE',
-      headers: {
-        'Content-Type': 'application/json',
-        'Accept': 'application/json'
-      },
-    })
-    .then(res => res.json())
-    .then(data => {
-      console.log(data);
-      dispatch({type: 'DELETEPLANT'})
-    })
+    const token = localStorage.getItem('token')
+    if(token){
+      fetch(`http://localhost:3000/api/v1/plants/${plant.id}`, {
+        method: 'DELETE',
+        headers: {
+          'Authorization': `Bearer ${token}`,
+          'Content-Type': 'application/json',
+          'Accept': 'application/json'
+        },
+      })
+      .then(res => res.json())
+      .then(data => {
+        // console.log(data);
+        dispatch({type: 'DELETEPLANT', payload: data})
+      })
+    }
   }
 }
